Restructure client folders
