--- client/src/components/Form/Form.js
@@ -1,8 +1,0 @@
-import React from "react";
-import { Form } from "./styles";
-
-export default () => {
-  return <Form></Form>;
-};
-
-//probably needs to go this whole component
